+ RecipeRecord and its Comparator (unfinished. Waiting on SQL)
Home fragment completely redone to accomodate a RecipeRecord
class. This was done to allow for easier sorting since I had no
idea how to sort 3 parallel ArrayLists. Plus it allows for the
original bookmark link to persist just in case we need it for
building a Meal viewing page. Can be updated to hold more data easily.