Working search fragment (searching works. Rest is still being worked on)
Built the backend for the Search Fragment. It now can search the API.

APIHandler was modified to allow for two separate courses of action
depending on whether it was asked to carry out a search or bookmarked
recipe retrieval.

In MainActivity the sendPreferences function was removed because it was
made redundant.